+++ Kventin.Backend/Kventin.Services/Dtos/Schedule/ReturnScheduleDto.cs
@@ -5,7 +5,7 @@ namespace Kventin.Services.Dtos.Schedule
     public class ReturnScheduleDto
     {
         [Required]
-        public required int ScheduleId { get; set; }
+        public required long ScheduleId { get; set; }
         public List<ReturnScheduleItemDto> ScheduleItems { get; set; } = [];
     }
 }
