+++ Kventin.Backend/Kventin.Services/Dtos/Schedule/ReturnScheduleItemDto.cs
@@ -6,7 +6,7 @@ namespace Kventin.Services.Dtos.Schedule
     public class ReturnScheduleItemDto
     {
         [Required]
-        public int ScheduleItemId { get; set; }
+        public long ScheduleItemId { get; set; }
 
         [Required]
         public required string DayOfWeek { get; set; }
