+++ Kventin.Backend/Kventin.Services/Interfaces/Services/ISubjectService.cs
@@ -5,9 +5,9 @@ namespace Kventin.Services.Interfaces.Services
     public interface ISubjectService
     {
         public Task CreateSubject(string subjectName);
-        public Task<List<SubjectDto>> GetAllSubjects(int userId, List<string> userRoles, int childId);
-        public Task<SubjectDto> GetSubjectByid(int id);
-        public Task DeleteSubjectById(int id);
-        public Task UpdateSubjectById(int id, string newSubjectName);
+        public Task<List<SubjectDto>> GetAllSubjects(long userId, List<string> userRoles, long childId);
+        public Task<SubjectDto> GetSubjectByid(long id);
+        public Task DeleteSubjectById(long id);
+        public Task UpdateSubjectById(long id, string newSubjectName);
     }
 }
